+++ server/handlers/helpers.go
@@ -1,7 +1,7 @@
-package main
+package handlers
 
 // Helper function to check if a string is in a slice of strings
-func contains2(slice []string, str string) bool {
+func contains(slice []string, str string) bool {
 	for _, s := range slice {
 		if s == str {
 			return true
